Overhaul of Question 2 Class
The third sample output for the second question did not correspond to the output shown in the terminal. The solution is to change the switch case from month -> market to market -> month, however this change in logic makes the code look significantly uglier. The output now corresponds to what is shown in the sample outputs. Documentation is left.